+++ src/queuemanager/PriorityItem.java
@@ -1,5 +1,7 @@
 package queuemanager;
 
+import java.util.Date;
+
 /**
  * A wrapper for bundling up an item and its integer priority.
  * 
@@ -8,9 +10,9 @@ package queuemanager;
 public class PriorityItem<T> {
 
     private final T item;
-    private final int priority;
+    private final long priority;
 
-    public PriorityItem(T item, int priority) {
+    public PriorityItem(T item, long priority) {
         this.item = item;
         this.priority = priority;
     }
@@ -19,7 +21,7 @@ public class PriorityItem<T> {
         return item;
     }
 
-    public int getPriority() {
+    public long getPriority() {
         return priority;
     }
